+++ app/main/__init__.py
@@ -2,4 +2,5 @@ from flask import Blueprint
 
 main = Blueprint('main', __name__)
 
-from . import views, errors, suppliers
+from . import errors
+from .views import crown_hosting, digital_services_framework, g_cloud, marketplace, suppliers
