+++ app/main/views/g_cloud.py
@@ -1,72 +1,29 @@
-# coding=utf-8
-
+# -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-from datetime import datetime
-from flask import abort, render_template, request, redirect, \
-    url_for, current_app
+from flask import abort, render_template, request, redirect, current_app
 
-from dmutils.deprecation import deprecated
 from dmutils.formats import get_label_for_lot_param, dateformat
-from dmapiclient import HTTPError, APIError
+from dmapiclient import HTTPError
 from dmutils.formats import LOTS
-from dmutils.content_loader import ContentNotFoundError
 
-from . import main
-from ..presenters.search_presenters import (
+from ...main import main
+from ...presenters.search_presenters import (
     filters_for_lot,
     set_filter_states,
 )
-from ..presenters.search_results import SearchResults
-from ..presenters.search_summary import SearchSummary
-from ..presenters.service_presenters import Service
-from ..helpers.shared_helpers import get_one_framework_by_status_in_order_of_preference
-from ..helpers.search_helpers import (
+from ...presenters.search_results import SearchResults
+from ...presenters.search_summary import SearchSummary
+from ...presenters.service_presenters import Service
+from ...helpers.search_helpers import (
     get_keywords_from_request, get_template_data, pagination,
     get_page_from_request, query_args_for_pagination,
     get_lot_from_request, build_search_query,
     clean_request_args
 )
 
-from ..exceptions import AuthException
-from .. import search_api_client, data_api_client, content_loader
-
-
-@main.route('/')
-def index():
-    template_data = get_template_data(main, {})
-    temporary_message = {}
-
-    try:
-        frameworks = data_api_client.find_frameworks().get('frameworks')
-        framework = get_one_framework_by_status_in_order_of_preference(
-            frameworks,
-            ['open', 'coming', 'pending']
-        )
-
-        if framework is not None:
-            content_loader.load_messages(framework.get('slug'), ['homepage-sidebar'])
-            temporary_message = content_loader.get_message(
-                framework.get('slug'),
-                'homepage-sidebar',
-                framework.get('status')
-            )
-
-    # if no framework is found (should never happen), ditch the message and load the page
-    except APIError:
-        pass
-    # if no message file is found (should never happen), throw a 500
-    except ContentNotFoundError:
-        current_app.logger.error(
-            "contentloader.fail No message file found for framework. "
-            "framework {} status {}".format(framework.get('slug'), framework.get('status')))
-        abort(500)
-
-    return render_template(
-        'index.html',
-        temporary_message=temporary_message,
-        **template_data
-    )
+from ...exceptions import AuthException
+from app import search_api_client, data_api_client, content_loader
 
 
 @main.route('/g-cloud')
@@ -79,28 +36,6 @@ def index_g_cloud():
 def framework_g_cloud():
     template_data = get_template_data(main, {})
     return render_template('content/framework-g-cloud.html', **template_data)
-
-
-@main.route('/digital-services/framework')
-def framework_digital_services():
-    template_data = get_template_data(main, {})
-    return render_template(
-        'content/framework-digital-services.html', **template_data
-    )
-
-
-@main.route('/crown-hosting')
-def index_crown_hosting():
-    template_data = get_template_data(main, {})
-    return render_template('content/index-crown-hosting.html', **template_data)
-
-
-@main.route('/crown-hosting/framework')
-def framework_crown_hosting():
-    template_data = get_template_data(main, {})
-    return render_template(
-        'content/framework-crown-hosting.html', **template_data
-    )
 
 
 @main.route('/buyers-guide')
@@ -127,32 +62,6 @@ def suppliers_guide_g_cloud():
     template_data = get_template_data(main, {})
     return render_template(
         'content/suppliers-guide-g-cloud.html', **template_data
-    )
-
-
-@main.route('/cookies')
-def cookies():
-    template_data = get_template_data(main, {})
-    return render_template(
-        'content/cookies.html', **template_data
-    )
-
-
-@main.route('/terms-and-conditions')
-def terms_and_conditions():
-    template_data = get_template_data(main, {})
-    return render_template(
-        'content/terms-and-conditions.html', **template_data
-    )
-
-
-# support legacy urls generated by the grails app
-@main.route('/service/<service_id>')
-@deprecated(dies_at=datetime(2016, 1, 1))
-def redirect_service_page(service_id):
-    return redirect(url_for(
-        ".get_service_by_id",
-        service_id=service_id.upper()), code=301
     )
 
 
@@ -211,12 +120,6 @@ def get_service_by_id(service_id):
         abort(404, "Service ID '%s' can not be found" % service_id)
     except HTTPError as e:
         abort(e.status_code)
-
-
-@main.route('/search')
-@deprecated(dies_at=datetime(2016, 1, 1))
-def redirect_search():
-    return redirect(url_for(".search", **request.args), code=301)
 
 
 @main.route('/g-cloud/search')
