+++ tests/app/views/test_errors.py
@@ -6,7 +6,7 @@ from dmapiclient import HTTPError
 from ...helpers import BaseApplicationTest
 
 
-@mock.patch('app.main.views.search_api_client')
+@mock.patch('app.main.views.g_cloud.search_api_client')
 class TestErrors(BaseApplicationTest):
     def test_404(self, search_api_mock):
         res = self.client.get('/g-cloud/service/1234')
